refactor(raisers): change loggers to raisers
re #7
